+++ docx_json_replacer/__init__.py
@@ -1,5 +1,5 @@
 from .docx_replacer import DocxReplacer, replace_docx_template
 from .table_handler import TableHandler
 
-__version__ = "0.5.1"
+__version__ = "0.7.1"
 __all__ = ["DocxReplacer", "replace_docx_template", "TableHandler"]